Mission LadderGame version1 what is unfinished
